backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
pydantic
sqlmodel
segno
jinja2
bcrypt
pwdlib[argon2]
PyJWT
python-multipart